+++ api/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_01_01_000005) do
+ActiveRecord::Schema[8.1].define(version: 2026_01_01_000006) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -95,6 +95,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_01_01_000005) do
     t.string "branch_name"
     t.datetime "created_at", null: false
     t.bigint "host_id"
+    t.datetime "last_activity_at"
     t.string "mode", default: "review", null: false
     t.text "objective"
     t.string "repository_path"